+++ Conclusion.java
@@ -1,0 +1,12 @@
+public class Conclusion {
+    public static void conclude() {
+        System.out.println("=".repeat(100));
+        System.out.println("Developed by: ");
+        System.out.println("🍺٩(˘◡˘ )" + " ".repeat(10) + " Tran Nguyen Anh Khoa");
+        System.out.println("Hua Minh Thu" + " ".repeat(10) + " (✿◠‿◠) ");
+        System.out.println(" (•◡•) /" + " ".repeat(10) + " Tran Phong");
+
+        System.out.println("=".repeat(100));
+        System.out.println("Program ended!");
+    }
+}
