+++ README.md
@@ -1,1 +1,27 @@
 # Programming1_Asignment
+
+TEAM CONTRIBUTION
+Tran Nguyen Anh Khoa(s3863956):0
+-
+-
+GIT HUB REPO
+https://github.com/KultrasHub/Programming1_Asignment.git
+YOUTUBE LINK
+
+FEATURE DISCLAIMER:
+4/22: 200
+4/23: 212
+4/24: 212
+4/25: 212
+4/26: 220
+if 4/23 to 4/26 get into a group and 4/22 is not in your chosen time range
+the total value of the group will be (220-200) rather than (220-212) eventhough 4/22 is not in the group.
+--
+We made this feature to prevent the scenarios where
+4/22: 200
+4/23: 212
+4/24: 212
+4/25: 212
+4/26: 212
+and if 4/23 to 4/26 belongs to a group, the total value of this group is 0.
+--
